backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.7.0
openai>=1.0.0
google-cloud-texttospeech>=2.16.0
requests>=2.31.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
